+++ tests/ai_safety/evalhub/utils.py
@@ -1,11 +1,12 @@
 import socket
-from typing import Final
+from typing import Any, Final
 
 import requests
 import structlog
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.config_map import ConfigMap
 from ocp_resources.evalhub import EvalHub
+from ocp_resources.mlflow import MLflow
 from ocp_resources.job import Job
 from ocp_resources.role_binding import RoleBinding
 from ocp_resources.service_account import ServiceAccount
@@ -37,6 +38,19 @@ from utilities.guardrails import get_auth_headers
 from utilities.kueue_utils import Workload
 
 LOGGER = structlog.get_logger(name=__name__)
+
+
+class MLflowWithWorkspaces(MLflow):
+    """MLflow CR with workspaceLabelSelector support."""
+
+    def __init__(self, workspace_label_selector: dict[str, Any] | None = None, **kwargs: Any) -> None:
+        super().__init__(**kwargs)
+        self._workspace_label_selector = workspace_label_selector
+
+    def to_dict(self) -> None:
+        super().to_dict()
+        if self._workspace_label_selector is not None and "spec" in self.res:
+            self.res["spec"]["workspaceLabelSelector"] = self._workspace_label_selector
 
 
 class TransientEvalhubHealthError(Exception):
